requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-community
python-dotenv
google-generativeai
faiss-cpu
